Added error handling to redis jobs:
We follow an exceptions as values approach for job error handling now.

This commit adds an decorator which wraps any job function and allows to
returns any raised exception as a value. These are than stored in the redis
store.
In our session status endpoint we now additionally check a jobs results
and if it is of exception type we set the status as failed.

Remark: Using the rq failed job queue does not work for this as they do
not serialize the raised exceptions properly only the traceback as str.
No idea why they think this is a proper way on how to handle exceptions...